Refactored code to improve performance. Replaced all Double types by Float. This should yield a better performance in most Accelerate functions while accuracy is only slightly worse

+++ IndoorLocation/IndoorLocation/BayesianFilter.swift
@@ -10,7 +10,7 @@ import Accelerate
 
 class BayesianFilter {
     
-    func computeAlgorithm(measurements: [Double], successCallback: @escaping (CGPoint) -> Void) {
+    func computeAlgorithm(measurements: [Float], successCallback: @escaping (CGPoint) -> Void) {
         // Least squares algorithm:
         guard let anchors = IndoorLocationManager.shared.anchors else {
             fatalError("No anchors found!")
